brain_core/requirements.txt
grpcio==1.78.1
grpcio-tools==1.78.1
protobuf==6.33.5
